Override ARP when sending unicast WoL
This code creates a system-wide, permanent (but not persistent - i.e.
it won't survive reboots) ARP entry for the TV IP address before sending
unicast ("device IP") WoL packets. The goal is to keep WoL working even
if the TV is ignoring Ethernet broadcast packets, including ARP
request.

In an effort to avoid making permanent changes to system-wide network
configuration (which would be user-hostile), we delete the ARP entry we
just added immediately after the packet is sent.

Fixes #19.